pyproject.toml
[project]
name = "strava-free"
version = "1.0.0"
description = "Simple Strava activity fetcher"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv>=1.0.0",
    "httpx>=0.24.0"
]

